feat: Implement user statistics display with pie chart
This commit introduces a comprehensive user statistics screen (`UserStatsFragment`) that displays detailed profile information and anime list statistics using a pie chart and summary cards.

Key changes:

- **User Statistics UI (`UserStatsFragment.xml`):**
    - Redesigned the layout to be a `NestedScrollView` containing user profile details and anime statistics.
    - Added a `PieChart` (from `MPAndroidChart`) to visualize the distribution of anime by status (Watching, Completed, Dropped, etc.).
    - Introduced `include_legend_status_watching.xml` and `include_summary_stats.xml` for reusable UI components to display chart legends and summary data (total episodes, rewatched count, mean score).
    - Added a `LinearProgressIndicator` to show loading state.
- **User Statistics Logic (`UserStatsFragment.kt`):**
    - Refactored to observe `isLoading` and `userInfo` from `UserStatsViewModel`.
    - Implemented methods to populate user profile info, anime statistics, and configure the `PieChart`.
    - Created `UiUtils.createStatSpannable` to format text with different sizes for stat values and their labels.
    - Added `StatusInfo.kt` data class to manage status-related data for the chart and legends.
- **ViewModel (`UserStatsViewModel.kt`):**
    - Introduced an `isLoading` `StateFlow` to manage the visibility of the loading indicator.
    - Fetching user stats now correctly sets the loading state before and after the API call.
- **API & Data Models:**
    - `UserInfoRepositoryImpl.kt`: Updated the API call to include `USER_INFO_ANIME_STATS_FIELDS` to fetch anime statistics.
    - `UserAnimeStatistics.kt`: Changed data types for `numDays...` and `meanScore` from `Double` to `Float` to match the API response.
- **Dependencies & Build:**
    - Added the `MPAndroidChart` library (`com.github.PhilJay:MPAndroidChart`) to display the pie chart.
    - Added the JitPack repository in `settings.gradle.kts`.
    - Incremented `versionCode` to 34 and `versionName` to "0.2.5".
- **String Resources (`strings.xml`):**
    - Added new strings for the statistics screen, such as `label_mean_score` and `label_rewatched`.
- **UI Refinements:**
    - `ProfileActivity.kt`: Refactored toolbar handling for better logic separation and clarity when navigating between fragments.
    - Removed `layout_scrollFlags` from the `Toolbar` in `activity_profile.xml`.

+++ app/src/main/java/com/animegatari/hayanime/data/model/UserAnimeStatistics.kt
@@ -23,22 +23,22 @@ data class UserAnimeStatistics(
     val numItems: Int? = null,
 
     @field:SerializedName("num_days_watched")
-    val numDaysWatched: Double? = null,
+    val numDaysWatched: Float? = null,
 
     @field:SerializedName("num_days_watching")
-    val numDaysWatching: Double? = null,
+    val numDaysWatching: Float? = null,
 
     @field:SerializedName("num_days_completed")
-    val numDaysCompleted: Double? = null,
+    val numDaysCompleted: Float? = null,
 
     @field:SerializedName("num_days_on_hold")
-    val numDaysOnHold: Double? = null,
+    val numDaysOnHold: Float? = null,
 
     @field:SerializedName("num_days_dropped")
-    val numDaysDropped: Double? = null,
+    val numDaysDropped: Float? = null,
 
     @field:SerializedName("num_days")
-    val numDays: Double? = null,
+    val numDays: Float? = null,
 
     @field:SerializedName("num_episodes")
     val numEpisodes: Int? = null,
@@ -47,5 +47,5 @@ data class UserAnimeStatistics(
     val numTimesRewatched: Int? = null,
 
     @field:SerializedName("mean_score")
-    val meanScore: Double? = null,
+    val meanScore: Float? = null,
 )
